Refine snapshotter module
