+++ pkg/backup/snapshotter/snapshotter_gcp.go
@@ -1,0 +1,68 @@
+// Copyright 2022 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package snapshotter
+
+import (
+	"fmt"
+	"regexp"
+	"strings"
+
+	"github.com/pingcap/tidb-operator/pkg/apis/pingcap/v1alpha1"
+	"github.com/pingcap/tidb-operator/pkg/backup/constants"
+	"github.com/pingcap/tidb-operator/pkg/controller"
+	corev1 "k8s.io/api/core/v1"
+)
+
+type GCPSnapshotter struct {
+	BaseSnapshotter
+}
+
+func (s *GCPSnapshotter) Init(deps *controller.Dependencies, conf map[string]string) error {
+	s.BaseSnapshotter.Init(deps, conf)
+	s.volRegexp = regexp.MustCompile(`^projects\/[^\/]+\/(zones|regions)\/[^\/]+\/disks\/[^\/]+$`)
+	return nil
+}
+
+func (s *GCPSnapshotter) GetVolumeID(pv *corev1.PersistentVolume) (string, error) {
+	if pv == nil {
+		return "", nil
+	}
+
+	if pv.Spec.CSI != nil {
+		driver := pv.Spec.CSI.Driver
+		if driver == constants.PdCSIDriver {
+			handle := pv.Spec.CSI.VolumeHandle
+			if !s.volRegexp.MatchString(handle) {
+				return "", fmt.Errorf("invalid volumeHandle for CSI driver:%s, expected projects/{project}/zones/{zone}/disks/{name}, got %s",
+					constants.PdCSIDriver, handle)
+			}
+			l := strings.Split(handle, "/")
+			return l[len(l)-1], nil
+		}
+		return "", fmt.Errorf("unable to handle CSI driver: %s", driver)
+	}
+
+	if pv.Spec.GCEPersistentDisk != nil {
+		if pv.Spec.GCEPersistentDisk.PDName == "" {
+			return "", fmt.Errorf("spec.gcePersistentDisk.pdName not found")
+		}
+		return pv.Spec.GCEPersistentDisk.PDName, nil
+	}
+
+	return "", nil
+}
+
+func (s *GCPSnapshotter) PrepareBackupMetadata(b *v1alpha1.Backup, tc *v1alpha1.TidbCluster, ns string) (string, error) {
+	return s.BaseSnapshotter.prepareBackupMetadata(b, tc, ns, s)
+}
